% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/avg_by_interp.R
\name{avg_by_fdp}
\alias{avg_by_fdp}
\title{Get average tpr and fdp rates over replicates}
\usage{
avg_by_fdp(tpr.list, fdp.list, npoints = 200)
}
\arguments{
\item{tpr.list}{A list of vectors giving true positive rates}

\item{npoints}{Number of points to evaluate}

\item{fpr.list}{A list of vectors giving false positive rates}
}
\value{
list with fpr, tpr and s.e
}
\description{
Get average tpr and fdp rates over replicates
}

