% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_clusters.R
\name{count_clusters_merged}
\alias{count_clusters_merged}
\title{Count clusters}
\usage{
count_clusters_merged(x, z, z0, seg.ends = NULL, except = NULL)
}
\arguments{
\item{x}{Data (already smoothed)}

\item{z}{Threshold: data frame k by nseg + 1}

\item{z0}{threshold for merging - separate regions at level}

\item{except}{Intervals object of area to remove
z that are subsets of a single region at level z0 will be merged.}
}
\value{
A vector the length of z giving the number of clusters at each level
}
\description{
Count clusters  given a threshold and level for merging
}

