% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/discopony.R
\name{discopony_maxes1}
\alias{discopony_maxes1}
\title{Get peak heights and permuted peak heights from dnase 1 data}
\usage{
discopony_maxes1(dat.file, pheno.file, s0, zmin, seed, n.perm, strt = NULL,
  stp = NULL, out.file = NULL, max1.file = NULL, z0 = zmin * 0.3,
  bandwidth = 50, maxit = 50)
}
\arguments{
\item{dat.file}{Data file. Should have first
column as position and a header giving sample names. One row per base pair.}

\item{pheno.file}{Phenotype file. First column should correspond to the header of dat.file.}

\item{s0}{Variance inflation constant.}

\item{zmin}{Lower bound on significance thresholds.}

\item{seed}{Set a seed for running permutations (Required)}

\item{n.perm}{Number of permutations}

\item{strt}{Region start base pair (may not be the same as where dat.file starts)}

\item{stp}{Region stop base pair.}

\item{out.file}{Output file (optional)}

\item{max1.file}{File with data from running with n.perm=0}

\item{z0}{Reference level for merging}

\item{bandwidth}{Smoother bandwidth.}

\item{maxit}{Maximum iterations for rlm.}

\item{z0}{Reference level for merging}
}
\value{
Saves an RData file with maxes list. Returns the length of that list.
}
\description{
Finds peak heights and permuatation peak heights using dnase 1 data.
}

