% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_clusters2.R
\name{get_clusters2}
\alias{get_clusters2}
\title{Count clusters at several levels of z and set threshold}
\usage{
get_clusters2(smoothed.stats, pos, zmin, z0 = 0.3 * zmin, level = c(0.02,
  0.05, 0.1, 0.2), segment.bounds = NULL)
}
\arguments{
\item{smoothed.stats}{Matrix of smoothed stats.}

\item{pos}{Length p vector of positions}

\item{zmin}{Vector of length 1 or 2}

\item{level}{Threshold at which to control FDR.}

\item{segment.bounds}{matrix with two columns giving segment bounds}
}
\value{
A list with elements:
\describe{
\item{\code{clust_num}}{Matrix of cluster counts nz x (1+n.perm)}
\item{\code{bw}}{ Bandwidth used for smoothing}
\item{\code{clust}}{ Clusters at zsel }
\item{\code{z}}{ All levels looked at}
\item{\code{z0}}{ Reference level used for merging}
\item{\code{fdr}}{ Estimated fdr at each level}
\item{\code{zsel}}{ Selected level of z}
}
}
\description{
Count clusters at several levels of z and set threshold. This function is mostly useful
for simulations when all of the smoothed statistics are in one data frame. For larger problems see the vignette.
}

