% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_stats.R
\name{get_stats_huber}
\alias{get_stats_huber}
\title{Get Huber stats}
\usage{
get_stats_huber(dat, labs, perm_labs, s0 = 0.1)
}
\arguments{
\item{dat}{Matrix of data (n x p)}

\item{labs}{Labels (0s and 1s)}

\item{perm_labs}{Permutation labels (n x n.perm)}

\item{s0}{Additional variance}
}
\value{
A matrix p x (1 + n.perm)
The first column is for te original labels.
The next columns are for the permutation labels
}
\description{
Calculate Huber statistic at each position
}

