% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_stats.R
\name{huber_helper}
\alias{huber_helper}
\title{Just like huber_stats2 but returns coefficient and sd estimates}
\usage{
huber_helper(Y, labs, k = 1.345, maxit = 20)
}
\arguments{
\item{Y}{matrix (p x n)}

\item{labs}{0s and 1s}

\item{k}{Threshold for huber estimator in multiples of scale parameter.}

\item{maxit}{Maixum iterations to pass to rlm.}

\item{s0}{Additional variance}
}
\value{
2 by p matrix. Top row is coefficient estimate. Bottom row is sd estimates.
}
\description{
Calculate two sample Huber
}

