% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/huber_var.R
\name{huber_stats}
\alias{huber_stats}
\title{Calculate two sample Huber statsitic}
\usage{
huber_stats(Y, labs, k = 1.345, scale = NULL, s0 = 0)
}
\arguments{
\item{Y}{matrix (p x n)}

\item{labs}{0s and 1s}

\item{k}{Threshold for huber estimator in multiples of scale parameter.}

\item{scale}{Vector of length p giving scale parameter.
Will be calculated if not provided.}

\item{s0}{Additional variance}

\item{zero.val}{Value to use if total in one group is zero}
}
\value{
Vector of statistics length p
}
\description{
Calculate two sample Huber
}

