% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_stats.R
\name{huber_stats2}
\alias{huber_stats2}
\title{Calculate Huber statsitic for simple linear regression usnig rlm}
\usage{
huber_stats2(Y, labs, k = 1.345, s0 = 0, maxit = 20)
}
\arguments{
\item{Y}{matrix (p x n)}

\item{labs}{0s and 1s}

\item{k}{Threshold for huber estimator in multiples of scale parameter.}

\item{s0}{Additional variance}

\item{maxit}{Maixum iterations to pass to rlm.}
}
\value{
Vector of test statistics.
}
\description{
Calculate two sample Huber
}

