% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/huber_var.R
\name{huber_var}
\alias{huber_var}
\title{Calculate variance of the huber estimator}
\usage{
huber_var(x, muH, k)
}
\arguments{
\item{x}{Data}

\item{muH}{Estimate of true huber mean}

\item{k}{Absolute threshold}
}
\value{
variance
}
\description{
Calculate variance of the huber estimator using influence functions
}

