% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/huber_var.R
\name{huber_var_eq6.6}
\alias{huber_var_eq6.6}
\title{Calculate variance of the huber estimator using equation 6.6 from Huber.}
\usage{
huber_var_eq6.6(f)
}
\arguments{
\item{f}{rlm object}
}
\value{
variance
}
\description{
Calculate variance of the huber estimator using influence functions
}

