% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{lm_stats}
\alias{lm_stats}
\title{Calculate two sample linear regression test statistics}
\usage{
lm_stats(Y, x, s0)
}
\arguments{
\item{Y}{matrix of data (p x n)}

\item{x}{trait value (0s and 1s)}

\item{s0}{Additional variance}
}
\value{
3 by p matrix
}
\description{
Calculate linear regression test statsitic
}

