% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_clusters.R
\name{name_clusters_merged}
\alias{name_clusters_merged}
\title{Return clusters at a given level}
\usage{
name_clusters_merged(x, z, z0, labs = FALSE, sgn = FALSE)
}
\arguments{
\item{x}{Data (already smoothed)}

\item{z}{Threshold}

\item{z0}{threshold for merging - separate regions at level
z that are subsets of a single region at level z0 will be merged.}
}
\value{
Intervals object giving clusters
}
\description{
Find indices of clusters  given a threshold and level for merging
}

