% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pois_regression.R
\name{pois_regression}
\alias{pois_regression}
\title{Calculate two sample Poisson regression statsitic}
\usage{
pois_regression(Y, labs, s0, zero.val = 1e-11)
}
\arguments{
\item{Y}{matrix of data (p x n)}

\item{labs}{0s and 1s}

\item{s0}{Additional variance}

\item{zero.val}{Value to use if total in one group is zero}
}
\value{
A length p vector of statistics
}
\description{
Calculate two sample Poisson regression statsitic
}

