% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{qpois_stats_binary}
\alias{qpois_stats_binary}
\title{Calculate two sample Quasi-Poisson regression statistic (binary predictor)}
\usage{
qpois_stats_binary(Y, x, s0, zero.val = 1e-11)
}
\arguments{
\item{Y}{matrix of data (p x n)}

\item{x}{trait value (0s and 1s)}

\item{s0}{Additional variance}

\item{zero.val}{Value to use if total in one group is zero}
}
\value{
3 by p matrix
}
\description{
Calculate two sample Quasi-Poisson regression statsitic
}

