% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stats.R
\name{qpois_stats_continuous}
\alias{qpois_stats_continuous}
\title{Calculate two sample Poisson regression statsitic for any type of predictor}
\usage{
qpois_stats_continuous(Y, x, s0)
}
\arguments{
\item{Y}{matrix of data (p x n)}

\item{x}{trait value (0s and 1s)}

\item{s0}{Additional variance}
}
\value{
3 by p matrix
}
\description{
Calculate Quasi-Poisson regression statsitic
}

