% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/replicate_typeseq.R
\name{replicate_typeseq}
\alias{replicate_typeseq}
\title{Replicate one type sequence many times}
\usage{
replicate_typeseq(type.sequence, file = NULL, sample.size = c(20, 20),
  seed = NULL, n.rep = 500, s0 = c(0, 0, 0), type.def = NULL,
  stat.names = c("Poisson", "Huber", "t-test"))
}
\arguments{
\item{type.sequence}{Vector of integers in 1:6 giving the type of each subregion.}

\item{sample.size}{Vector of length 2 giving sample sizes}

\item{seed}{Optional - set a seed.}

\item{n.rep}{Number of simulations}

\item{s0}{Vector of length 3 giving additional variance to add to Poisson, Huber and t-test statistics.}

\item{n.perms}{Number of permutations to use to estimate lambda.}
}
\value{
A list of statistic matrices, one for each test type
}
\description{
Generate data from a type sequence to estimate z
}

