% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_cfdr_sims.R
\name{run_cfdr_sims}
\alias{run_cfdr_sims}
\title{Run Simulations}
\usage{
run_cfdr_sims(type.sequence, B = 10, sample.size = c(20, 20), seed = NULL,
  n.perms = 500, s0 = c(0.1, 0.1, 0), save.data = FALSE,
  type.def = NULL)
}
\arguments{
\item{type.sequence}{Vector of integers in 1:6 giving the type of each subregion.}

\item{B}{Number of simulations}

\item{sample.size}{Vector of length 2 giving sample sizes}

\item{seed}{Optional - set a seed.}

\item{n.perms}{Number of permutations to use to estimate lambda.}

\item{s0}{Vector of length 3 giving additional variance to add to Poisson, Huber and t-test statistics.}
}
\value{
A list with elements:
\describe{
\item{\code{n.false.discoveries}}{ Matrix 3 x B Giving total false discoveries for each method and each simulation.}
\item{\code{true.discoveries}}{ Array 3 x q x B where q is the number of regions of type 4:6}
}
}
\description{
Run simulations
}

