% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cfdr_sims.R
\name{run_cfdr_sims1}
\alias{run_cfdr_sims1}
\title{Run Simulations}
\usage{
run_cfdr_sims1(type.sequence, sample.size = c(20, 20), seed = NULL,
  n.perms = 500, s0 = c(0, 0, 0), level = c(0.02, 0.05, 0.1, 0.2),
  save.data = FALSE, type.def = NULL, huber.maxit = 50,
  stat.names = c("Poisson", "Huber", "t-test"), file.name = NULL)
}
\arguments{
\item{type.sequence}{Vector of integers in 1:6 giving the type of each subregion.}

\item{sample.size}{Vector of length 2 giving sample sizes}

\item{seed}{Optional - set a seed.}

\item{n.perms}{Number of permutations to use to estimate lambda.}

\item{s0}{Vector of length 3 giving additional variance to add to Poisson, Huber and t-test statistics.}

\item{B}{Number of simulations}
}
\value{
A list with elements:
\describe{
\item{\code{n.f.disc}}{ Number of false disccoveries}
\item{\code{n.t.disc}}{ Array 3 x q x B where q is the number of regions of type 4:6}
}
}
\description{
Run simulations
}

