% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_data.R
\name{sample_data}
\alias{sample_data}
\title{Sample data}
\usage{
sample_data(type.sequence, sample.size, type.def = NULL)
}
\arguments{
\item{type.sequence}{Vector of integers in 1:6 giving the type of each subregion.}

\item{sample.size}{Vector of length 2 giving sample sizes}
}
\value{
A list with elements:
\describe{
\item{\code{dat}}{ Matrix of observed data (200*w) x sum(sample.size) where w is length(type.sequence)}
\item{\code{means}}{ Matrix of sample specific profiles.}
}
}
\description{
Sample data given a sample size and a vector of types
}

