% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tpr_nfp.R
\name{tpr_nfp}
\alias{tpr_nfp}
\title{Calculate accuracy rates given a set of regions or a list of statsistics and a threshold}
\usage{
tpr_nfp(signal, discoveries = NULL, x = NULL, z = NULL, z0 = NULL)
}
\arguments{
\item{signal}{Intervals object giving regions of true signal}

\item{discoveries}{Intervals object giving discoveries}

\item{x}{vector of statistics. Provide with z and z0 in lieu of discoveries}

\item{z}{vector of thresholds. Should be length of x}

\item{z0}{reference level for merging}
}
\value{
A list with elements:
\describe{
\item{\code{tpr}}{ True positive rate}
\item{\code{nfp}}{ Number of false positives}
\item{\code{ntp}}{ Number of true positives}
\item{\code{fdp}}{ False discover proportion nfp/(nfp + ntp)}
}
}
\description{
Calculates tpr, number of true positives, number of false positives, and the false discovery proportion.
}

