% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/count_clusters.R
\name{count_clusters_merged}
\alias{count_clusters_merged}
\title{Count clusters}
\usage{
count_clusters_merged(x, z, z0)
}
\arguments{
\item{x}{Data (already smoothed)}

\item{z}{Threshold}

\item{z0}{threshold for merging - separate regions at level
z that are subsets of a single region at level z0 will be merged.}
}
\value{
A list with elements:
\describe{
\item{\code{nclust}}{ Total number of clusters.}
\item{\code{clust}}{ Intervals object giving clusters (merged).}
}
}
\description{
Count clusters  given a threshold and level for merging
}

