% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make_plot.R
\name{make_plot}
\alias{make_plot}
\title{Plot Simulations}
\usage{
make_plot(tot.rates, names, ltys, cols, shapes, levels = c(0.02, 0.05, 0.1,
  0.2))
}
\arguments{
\item{ltys}{Line type for each summary file}

\item{levels}{Levels in summ.files}

\item{summ.files}{List of summary files (see collect.R)}

\item{lty.legend}{Legend labels for line type}

\item{ymax}{Optional max height of FDR y axis}
}
\value{
nothing
}
\description{
Plot simulations
}

