% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/window_test.R
\name{window_test}
\alias{window_test}
\title{Test simulations in windows}
\usage{
window_test(windows, dat, pos, x, signal, huber.maxit = 50, s0 = c(0, 0, 0),
  level = c(0.02, 0.05, 0.1, 0.2))
}
\arguments{
\item{windows}{q by 2 matrix giving region boundaries}

\item{dat}{p x n matrix of data}

\item{pos}{Positions}

\item{x}{treatment status of each sample (length n)}

\item{signal}{}

\item{s0}{}

\item{level}{}
}
\value{
A list
}
\description{
Calculate test statistics binning over aupplied regions
}

