% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/window_test.R
\name{window_test}
\alias{window_test}
\title{Test simulations in windows}
\usage{
window_test(windows, type.sequence, dat, labs, s0 = c(0, 0, 0),
  stat.names = c("Poisson", "Huber", "t-test"))
}
\arguments{
\item{windows}{q by 2 matrix giving region boundaries}

\item{type.sequence}{Vector of integers in 1:6 giving the type of each subregion.}

\item{dat}{p x n matrix of data}

\item{labs}{0/1 treatment status of each sample (length n)}

\item{s0}{}

\item{stat.names}{Length k. Contains elements of "Poisson", "Huber", "t-test"}
}
\value{
A list with elements:
\describe{
\item{\code{stats}}{ matrix k by q giving statistic for each bin }
\item{\code{rate_list}}{list of length k. Each element is a three column matrix with columns fdp, tp, fp}
}
}
\description{
Calculate test statistics binning over aupplied regions
}

